restructure domain folders
