--- src/main/scala/krs/UserRepository.scala
@@ -1,6 +1,0 @@
-package krs
-
-trait UserRepository {
-  def loadUsers(): List[User]
-  def getUser(id: Int): Option[User]
-}
